Former-commit-id: 513ea0a4ef1357cdc5907eec45f80d68c4f01110

+++ Sources/Internal/Render/Shader.cpp
@@ -61,7 +61,7 @@ Shader::Shader()
     attributeNames = 0;
     activeAttributes = 0;
     activeUniforms = 0;
-    activeAttributesMask = 0;
+    requiredVertexFormat = 0;
     
     //uniforms = 0;
     uniformData = NULL;
@@ -279,7 +279,7 @@ Shader::~Shader()
     ReleaseShaderData();
 }
 
-void Shader::ReleaseShaderData()
+void Shader::ReleaseShaderData(bool deleteShader/* = true*/)
 {
     SafeDeleteArray(attributeNames);
     //SafeDeleteArray(uniforms);
@@ -289,7 +289,8 @@ void Shader::ReleaseShaderData()
     SafeRelease(vertexShaderData);
     SafeRelease(fragmentShaderData);
     
-    DeleteShaders();
+    if (deleteShader)
+        DeleteShaders();
 
     activeAttributes = 0;
     activeUniforms = 0;
@@ -351,7 +352,9 @@ void Shader::RecompileInternal(BaseObject * caller, void * param, void *callerDa
     char attributeName[512];
     DVASSERT(attributeNames == NULL);
     attributeNames = new FastName[activeAttributes];
-    activeAttributesMask = 0;
+
+    requiredVertexFormat = 0;
+
     for (int32 k = 0; k < activeAttributes; ++k)
     {
         GLint size;
@@ -364,11 +367,7 @@ void Shader::RecompileInternal(BaseObject * caller, void * param, void *callerDa
         {
             int32 attributeLocationIndex = glGetAttribLocation(program, attributeName);
             vertexFormatAttribIndeces[flagIndex] = attributeLocationIndex;
-            
-            if(attributeLocationIndex != -1)
-            {
-                activeAttributesMask |= (1 << flagIndex);
-            }
+            requiredVertexFormat |= 1<<flagIndex;
         }
     }
     
@@ -1348,6 +1347,8 @@ void Shader::Lost()
     program = 0;
     activeAttributes = 0;
     activeUniforms = 0;
+    
+    ReleaseShaderData(false);
 }
 
 void Shader::Invalidate()
